[220225] Modify color width struct (#38)
* [220225] modify : width float -> CGFloat로 변경

* [220225] modify : struct 를 CGFloat으로 변경해 CGfloat() 제거

+++ RxSwift/RxSwiftTIL/RxSwiftTIL/Model/ColorWidth.swift
@@ -9,6 +9,6 @@ import UIKit
 
 struct ColorWidth {
     let color : UIColor
-    let width : Float
+    let width : CGFloat
 }
